+++ validator/src/main/java/org/mustangproject/validator/PDFValidator.java
@@ -199,7 +199,7 @@ public class PDFValidator extends Validator {
 
 			boolean documentTypeValid=false;
 			for (int i = 0; i < nodes.getLength(); i++) {
-				if (nodes.item(i).getTextContent().equals("INVOICE")) {
+				if (nodes.item(i).getTextContent().equals("INVOICE")||nodes.item(i).getTextContent().equals("ORDER")) {
 					documentTypeValid=true;
 				}
 			}
@@ -220,7 +220,7 @@ public class PDFValidator extends Validator {
 			}
 			boolean documentFilenameValid=false;
 			for (int i = 0; i < nodes.getLength(); i++) {
-				String[] valueArray = { "factur-x.xml", "ZUGFeRD-invoice.xml", "zugferd-invoice.xml", "xrechnung.xml" };
+				String[] valueArray = { "factur-x.xml", "ZUGFeRD-invoice.xml", "zugferd-invoice.xml", "xrechnung.xml" , "order-x.xml" };
 				if (stringArrayContains(valueArray, nodes.item(i).getTextContent())) {
 					documentFilenameValid=true;
 				}