+++ History.md
@@ -1,10 +1,11 @@
 2.1.1
 =======
-2020-02-02
+2020-02-09
 
 - PR #217 seller order referenced document thanks to weclapp-dev
 - PR #218 allow recipient contact in XRechnung thanks to seeeeew
-- RE #221 remove outdated dependency thanks to heisej 
+- RE #221 remove outdated dependency thanks to heisej
+- RE #223 allow XRechnung without street thanks to murygin
 - only deploy library and validator to maven central, not core nor cli
 - a correction should reference the original invoice via invoiceReferencedDocument, not buyerOrderReferencedDocument
 - upgraded CEN Schematron validation (codes 24) to v1.3.4