+++ History.md
@@ -1,10 +1,15 @@
 2.1.1
 =======
+2020-02-09
 
-- PR #217 seller order referenced document
+- PR #217 seller order referenced document thanks to weclapp-dev
 - PR #218 allow recipient contact in XRechnung thanks to seeeeew
+- PR #221 remove outdated dependency thanks to heisej
+- PR #223 allow XRechnung without street thanks to murygin
 - only deploy library and validator to maven central, not core nor cli
 - a correction should reference the original invoice via invoiceReferencedDocument, not buyerOrderReferencedDocument
+- upgraded CEN Schematron validation (codes 24) to v1.3.4
+- added unit tests for ubl conversion and visualization
 
 2.1.0
 =======