+++ History.md
@@ -1,6 +1,7 @@
 - PR #225 exemption reasons only for certain tax category codes thanks to weclapp-dev
 - allow 1.2 and 2.0 in RDF versions for XRechnung 2.0 Referenzprofil 
 - also use shortcut "t" for extended in zf1 #230
+- falscher Text in Exception #237
 
 2.1.1
 =======