+++ History.md
@@ -1,3 +1,7 @@
+2.2.0
+=======
+2021-05-20
+
 - PR #225 exemption reasons only for certain tax category codes thanks to weclapp-dev
 - allow 1.2 and 2.0 in RDF versions for XRechnung 2.0 Referenzprofil 
 - also use shortcut "t" for extended in zf1 #230